Ensure directory ownership for abuild
In some circumstances, the bind mounts to the package build and output
directories can result in them being owned by root. This causes the
abuild process to fail when it attempts to create pkg/ and src/. By
explicitly changing ownership of these directories, the build process
will run as expected - regardless of the Docker environment.